requirements.txt
pandas
ta
requests
python-dotenv
pytz
